modify instruction list creator to handle single subj

+++ meg-distraction/create-instructions-list.py
@@ -4,12 +4,13 @@
 """ Generation the instructions files for each subject for mm-meg-distraction tasks
 """
 
+import sys
 import random
 import pandas as pd
 from pathlib import Path
+import argparse
 
 DEST_DIR = Path('instructions')
-N_SUBJECTS = 50
 N_RUNS = 28
 N_INSTRUCTIONS_PER_TYPE = 5
 INSTRUCTION_FILES = ['list.txt', 'math.txt', 'memory.txt']
@@ -44,20 +45,34 @@ def pick_instructions(tasks_df, n_instructions_per_type):
 
 
 if __name__ == '__main__':
-    tasks = merge_tasks_file_into_df(INSTRUCTION_FILES)
+    parser = argparse.ArgumentParser(description='Generate instructions for each run for a given participant')
+    parser.add_argument('--subject', type=int, required=True)  
+    args = parser.parse_args()
+
+    if args.subject is None:
+        parser.print_help()
+        sys.exit(1)
 
     DEST_DIR.mkdir(parents=True, exist_ok=True)
+    sub_file = DEST_DIR / f"sub-{args.subject}_instructions.tsv"
 
-    for s in range(1, N_SUBJECTS + 1):
-        sub_file = DEST_DIR / f"sub-{s}_instructions.tsv"    
-        x = pick_instructions(tasks, N_INSTRUCTIONS_PER_TYPE)
+    if sub_file.is_file():
+        resp = input(f'{sub_file} exists. Do you want to overwrite it? (y/n)')
+        if resp == 'n':
+            sys.exit(1)
+            
+    
+    tasks = merge_tasks_file_into_df(INSTRUCTION_FILES)
 
-        # insert no task runs
-        for i in range(len(x.index)):
-            x.loc[s % 2 + i - 0.5] = 'NoTask', 'NA'
-        x = x.sort_index().reset_index(drop=True)
-        x["run"] = x.index + 1
-        x[:N_RUNS].to_csv(sub_file, sep='\t')
+    x = pick_instructions(tasks, N_INSTRUCTIONS_PER_TYPE)
+
+    # insert no task runs
+    for i in range(len(x.index)):
+        x.loc[args.subject % 2 + i - 0.5] = 'NoTask', 'NA'
+    x = x.sort_index().reset_index(drop=True)
+    x["run"] = x.index + 1
+    
+    x[:N_RUNS].to_csv(sub_file, sep='\t')
